fix: repair proxy auth matcher, stale role guard, and duplicate scope schema
