+++ packages/auth/src/skill-runtime-capability.ts
@@ -12,15 +12,6 @@ const CAPABILITY_DIGEST_LENGTH = 43;
 const BASE64_URL_PATTERN = /^[A-Za-z0-9_-]+$/u;
 const TEXT_ENCODER = new TextEncoder();
 
-export const SkillRuntimeScopeSchema = z.enum([
-  "events:write",
-  "integrations:execute",
-  "skills:read",
-  "skills:write",
-]);
-
-export type SkillRuntimeScope = z.infer<typeof SkillRuntimeScopeSchema>;
-
 export interface MintedSkillRuntimeCapability {
   digest: string;
   expiresAt: number;
